chore(grid-snapping): extract shared parts into utility
