+++ lib/features/grid-snapping/GridSnapping.js
@@ -10,7 +10,10 @@ import {
   isNumber
 } from 'min-dash';
 
-export var SPACING = 10;
+import {
+  SPACING,
+  quantize
+} from './GridUtil';
 
 var LOWER_PRIORITY = 1200;
 var LOW_PRIORITY = 800;
@@ -294,12 +297,4 @@ function isNorth(direction) {
 
 function isWest(direction) {
   return direction.charAt(1) === 'w';
-}
-
-function quantize(value, quantum, fn) {
-  if (!fn) {
-    fn = 'round';
-  }
-
-  return Math[ fn ](value / quantum) * quantum;
 }
