+++ lib/features/grid-snapping/GridUtil.js
@@ -1,0 +1,9 @@
+export var SPACING = 10;
+
+export function quantize(value, quantum, fn) {
+  if (!fn) {
+    fn = 'round';
+  }
+
+  return Math[ fn ](value / quantum) * quantum;
+}
