+++ lib/features/grid-snapping/behavior/PasteBehavior.js
@@ -5,7 +5,7 @@ import CommandInterceptor from '../../../command/CommandInterceptor';
 import { getBBox } from '../../../util/Elements';
 import { getMid } from '../../../layout/LayoutUtil';
 
-import { SPACING } from '../GridSnapping';
+import { SPACING } from '../GridUtil';
 
 var HIGH_PRIORITY = 2000;
