+++ lib/features/grid-snapping/visuals/Grid.js
@@ -7,7 +7,7 @@ import {
 
 import { query as domQuery } from 'min-dom';
 
-import { SPACING } from '../GridSnapping';
+import { SPACING } from '../GridUtil';
 
 var GRID_COLOR = '#ccc',
     LAYER_NAME = 'djs-grid';
